+++ lib/backup/view/widgets/widgets.dart
@@ -1,0 +1,5 @@
+export 'backup_button.dart';
+export 'backups_list.dart';
+export 'sign_in_button.dart';
+export 'google_profile.dart';
+export 'confirm_dialog.dart';