+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [4.4.0] - 03-26-2025
+### Added
+- Auto backup to Google Drive
+- Restored backup overrides live database
+- Showing the app's current version in the drawer
+
 ## [4.3.0] - 03-22-2025
 ### Added
 - Google Drive backup